Refactor hook implementation
