+++ src/views/domain-page/domain-page-header-status-tag/domain-page-header-status-tag.tsx
@@ -1,5 +1,5 @@
 import DomainStatusTag from '@/views/shared/domain-status-tag/domain-status-tag';
-import useSuspenseDomainDescription from '@/views/shared/hooks/use-suspense-domain-description';
+import useSuspenseDomainDescription from '@/views/shared/hooks/use-domain-description/use-suspense-domain-description';
 
 import { type Props } from './domain-page-header-status-tag.types';
