+++ src/views/domain-page/domain-page-metadata/domain-page-metadata.types.ts
@@ -1,10 +1,5 @@
 import { type DomainDescription } from '../domain-page.types';
 
-export type UseSuspenseDomainPageMetadataParams = {
-  domain: string;
-  cluster: string;
-};
-
 export type DomainMetadata = {
   domainDescription: DomainDescription;
   isExtendedMetadataEnabled: boolean;
