+++ src/views/domain-page/hooks/use-is-extended-metadata-enabled/use-suspense-is-extended-metadata-enabled.ts
@@ -1,0 +1,7 @@
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+import getIsExtendedMetadataEnabledQueryOptions from './get-is-extended-metadata-enabled-query-options';
+
+export default function useSuspenseIsExtendedMetadataEnabled() {
+  return useSuspenseQuery(getIsExtendedMetadataEnabledQueryOptions());
+}
