+++ src/views/shared/hooks/use-domain-description/get-domain-description-query-options.ts
@@ -1,25 +1,25 @@
-import { useSuspenseQuery } from '@tanstack/react-query';
+import { type UseQueryOptions } from '@tanstack/react-query';
 
 import request from '@/utils/request';
 import { type RequestError } from '@/utils/request/request-error';
 import { type DomainDescription } from '@/views/domain-page/domain-page.types';
 
-import { type UseSuspenseDomainDescriptionProps } from './use-suspense-domain-description.types';
+import { type UseDomainDescriptionParams } from './use-domain-description.types';
 
-export default function useSuspenseDomainDescription({
+export default function getDomainDescriptionQueryOptions({
   domain,
   cluster,
-}: UseSuspenseDomainDescriptionProps) {
-  return useSuspenseQuery<
-    DomainDescription,
-    RequestError,
-    DomainDescription,
-    [string, { domain: string; cluster: string }]
-  >({
+}: UseDomainDescriptionParams): UseQueryOptions<
+  DomainDescription,
+  RequestError,
+  DomainDescription,
+  [string, { domain: string; cluster: string }]
+> {
+  return {
     queryKey: ['describeDomain', { domain, cluster }],
     queryFn: ({ queryKey: [_, params] }) =>
       request(`/api/domains/${params.domain}/${params.cluster}`).then((res) =>
         res.json()
       ),
-  });
+  };
 }
