+++ src/views/shared/hooks/use-domain-description/use-domain-description.types.ts
@@ -1,0 +1,4 @@
+export type UseDomainDescriptionParams = {
+  domain: string;
+  cluster: string;
+};
