--- src/views/shared/hooks/use-suspense-domain-description.types.ts
@@ -1,4 +1,0 @@
-export type UseSuspenseDomainDescriptionProps = {
-  domain: string;
-  cluster: string;
-};
